src/main/java/com/sockets/chatapp/model/User.java
package com.sockets.chatapp.model;

import org.springframework.data.annotation.Id;
import org.springframework.data.mongodb.core.mapping.Document;

@Document(collection = "users")
public class User {
    @Id
    private String id;
    private String username;
    private String password;

    // No-argument constructor
    public User() {
    }

    // All-arguments constructor
    public User(String id, String username, String password) {
        this.id = id;
        this.username = username;
        this.password = password;
    }

    // Getters
    public String getId() {
        return id;
    }

    public String getUsername() {
        return username;
    }

    // Normally you do not provide a password getter for security reasons
    // Provide this with caution
    public String getPassword() {
        return password;
    }

    // Setters
    public void setId(String id) {
        this.id = id;
    }

    public void setUsername(String username) {
        this.username = username;
    }

    public void setPassword(String password) {
        this.password = password;
    }

    // hashCode and equals (omitted for brevity)

    // toString (ensure not to expose sensitive data like password)
    @Override
    public String toString() {
        return "User{" +
                "id='" + id + '\'' +
                ", username='" + username + '\'' +
                // Do not print the password
                '}';
    }
}